try different split
