+++ src/replay/replayio.cc
@@ -1,0 +1,30 @@
+#include "src/execution/isolate-inl.h"
+#include "src/objects/string.h"
+#include "include/replayio.h"
+#include "src/replay/replayio.h"
+
+namespace v8 {
+namespace replayio {
+
+v8::internal::Handle<v8::internal::String> RecordReplayStringHandle(
+    const char* why, v8::internal::Isolate* isolate,
+    v8::internal::Handle<v8::internal::String> input) {
+  if (!v8::recordreplay::IsRecordingOrReplaying(why)) {
+    return input;
+  }
+  std::string str = input->ToCString().get();
+  v8::recordreplay::RecordReplayString(why, str);
+  return isolate->factory()->NewStringFromUtf8(base::CStrVector(str.c_str())).ToHandleChecked();
+}
+
+v8::internal::MaybeHandle<v8::internal::String> RecordReplayStringHandle(
+    const char* why, v8::internal::Isolate* isolate,
+    v8::internal::MaybeHandle<v8::internal::String> input) {
+  if (input.is_null()) {
+    return input;
+  }
+  return MaybeHandle<String>(RecordReplayStringHandle(why, isolate, input.ToHandleChecked()));
+}
+
+}  // namespace replayio
+}  // namespace v8
